@@ -1,14 +1,19 @@
 <?php
-namespace Technical\Unit\Yoanm\PhpUnitExtended\Listener;
+namespace Tests\Functional\Listener;
 
-use Prophecy\Argument;
+use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\Test;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestListener;
+use PHPUnit\Framework\TestSuite;
+use PHPUnit\Framework\Warning;
 use Prophecy\Prophecy\ObjectProphecy;
 use Yoanm\PhpUnitExtended\Listener\DelegatingListener;
 
 /**
  * @covers Yoanm\PhpUnitExtended\Listener\DelegatingListener
  */
-class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
+class DelegatingListenerTest extends TestCase
 {
     /** @var DelegatingListener */
     private $listener;
@@ -18,10 +23,10 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         $this->listener = new DelegatingListener();
     }
     
-    public function testListenerList()
+    public function testShouldManagerAListOfListener()
     {
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
 
         $this->listener->addListener($delegate->reveal());
 
@@ -31,14 +36,14 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function testAddError()
+    public function testShouldHandleError()
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
         /** @var \Exception|ObjectProphecy $exception */
         $exception = $this->prophesize(\Exception::class);
 
@@ -50,16 +55,35 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         $this->listener->addError($test->reveal(), $exception->reveal(), $time);
     }
 
-    public function testAddFailure()
+    public function testShouldHandleWarning()
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
-        /** @var \PHPUnit_Framework_AssertionFailedError|ObjectProphecy $exception */
-        $exception = $this->prophesize(\PHPUnit_Framework_AssertionFailedError::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
+        /** @var Warning|ObjectProphecy $exception */
+        $warning = $this->prophesize(Warning::class);
+
+        $this->listener->addListener($delegate->reveal());
+
+        $delegate->addWarning($test->reveal(), $warning->reveal(), $time)
+            ->shouldBeCalled();
+
+        $this->listener->addWarning($test->reveal(), $warning->reveal(), $time);
+    }
+
+    public function testShouldHandleFailure()
+    {
+        $time = 0.2;
+
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
+        /** @var AssertionFailedError|ObjectProphecy $exception */
+        $exception = $this->prophesize(AssertionFailedError::class);
 
         $this->listener->addListener($delegate->reveal());
 
@@ -69,14 +93,14 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         $this->listener->addFailure($test->reveal(), $exception->reveal(), $time);
     }
 
-    public function testAddIncompleteTest()
+    public function testShouldHandleIncompleteTest()
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
         /** @var \Exception|ObjectProphecy $exception */
         $exception = $this->prophesize(\Exception::class);
 
@@ -88,14 +112,14 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         $this->listener->addIncompleteTest($test->reveal(), $exception->reveal(), $time);
     }
 
-    public function testAddRiskyTest()
+    public function testShouldHandleRiskyTest()
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
         /** @var \Exception|ObjectProphecy $exception */
         $exception = $this->prophesize(\Exception::class);
 
@@ -107,14 +131,14 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
         $this->listener->addRiskyTest($test->reveal(), $exception->reveal(), $time);
     }
 
-    public function testAddSkippedTest()
+    public function testShouldHandleSkippedTest()
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
         /** @var \Exception|ObjectProphecy $exception */
         $exception = $this->prophesize(\Exception::class);
 
@@ -128,10 +152,10 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
 
     public function testStartTestSuite()
     {
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_TestSuite|ObjectProphecy $testSuite */
-        $testSuite = $this->prophesize(\PHPUnit_Framework_TestSuite::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var TestSuite|ObjectProphecy $testSuite */
+        $testSuite = $this->prophesize(TestSuite::class);
 
         $this->listener->addListener($delegate->reveal());
 
@@ -143,10 +167,10 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
 
     public function testEndTestSuite()
     {
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_TestSuite|ObjectProphecy $testSuite */
-        $testSuite = $this->prophesize(\PHPUnit_Framework_TestSuite::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var TestSuite|ObjectProphecy $testSuite */
+        $testSuite = $this->prophesize(TestSuite::class);
 
         $this->listener->addListener($delegate->reveal());
 
@@ -158,10 +182,10 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
 
     public function testStartTest()
     {
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
 
         $this->listener->addListener($delegate->reveal());
 
@@ -175,10 +199,10 @@ class DelegatingListenerTest extends \PHPUnit_Framework_TestCase
     {
         $time = 0.2;
 
-        /** @var \PHPUnit_Framework_TestListener|ObjectProphecy $delegate */
-        $delegate = $this->prophesize(\PHPUnit_Framework_TestListener::class);
-        /** @var \PHPUnit_Framework_Test|ObjectProphecy $test */
-        $test = $this->prophesize(\PHPUnit_Framework_Test::class);
+        /** @var TestListener|ObjectProphecy $delegate */
+        $delegate = $this->prophesize(TestListener::class);
+        /** @var Test|ObjectProphecy $test */
+        $test = $this->prophesize(Test::class);
 
         $this->listener->addListener($delegate->reveal());
 
